tpcc initial test and some other fixes

+++ system/global.h
@@ -285,8 +285,8 @@ extern UInt32 g_seq_thread_cnt;
 //MIXED_LOCK
 extern UInt32 g_calvin_thread_cnt; // number of calvin threads in the thread pool
 extern UInt64 g_data_shard_size; // number of tuples in each data shard
-extern uint16_t g_lower_bound;
-extern uint16_t g_upper_bound;
+extern UInt64 g_lower_bound;
+extern UInt64 g_upper_bound;
 extern UInt64 g_total_shard_num;
 extern UInt64 g_conflict_send_interval;
